pyproject.toml
[project]
name = "video-gen"
version = "0.1.0"
description = "AI TikTok video generation pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "typer>=0.9",
    "rich>=13.0",
    "anthropic>=0.40",
    "torch>=2.0",
    "diffusers>=0.31",
    "transformers>=4.40",
    "accelerate>=0.30",
    "sentencepiece>=0.2",
    "f5-tts>=1.0",
    "moviepy>=2.0",
    "Pillow>=10.0",
]

[project.scripts]
video-gen = "video_gen.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/video_gen"]
